Clean up how controllers respond to events
Controllers were responding to any event and just simply doing their action if the appropriate state was present, which is obviously not workable.

This change makes it so that the controllers depend on the Action log and only execute if the last action is something they're interested in.

This refactor can itself be moved out to Controller behavior definition files so that not all behaviors depend on the last action, just this particular "show the view" behavior.